+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Latest changes
 
+* Fix bug where `Request.url_for` is not working as intended [[yuval9313/FastApi-RESTful#90](https://github.com/yuval9313/FastApi-RESTful/issues/90)]
 * Update multiple dependencies using @dependebot
 * Fix `repeat_every` is only running once [#142](https://github.com/yuval9313/FastApi-RESTful/pull/142)